chore(awareness): split subscription modes into watchlist_modes.go, bump set_mutators baseline
The store crossed the 500-line gate; the mode vocabulary + IngestGlobs
are a coherent unit and move to their own file. set_mutators 111→112 is
deliberate: StateWatcher.SetFilter is genuine runtime mutation — the
entire point of #1192 is swapping the ingestion gate while running —
not construction-time wiring that belongs in a constructor.

Refs #1192.
