+++ scripts/architecture.baseline
@@ -1,5 +1,5 @@
 packages=64
 interfaces=83
 large_files=49
-set_mutators=111
+set_mutators=112
 database_opens=9
